Further tweak documentation workflow
Do not run documentation builds for the u/neophile branch, since
they won't have meaningful changes.  Build the documentation on pull
requests but skip the documentation upload, so we still check for
errors building the docs but don't try to create a new version for
every PR.  Ensure we run Docker builds on pull requests from ticket
branches in the same repository since we no longer run the workflow
on a push to such a branch.  Be explicit about running the CI
workflow on all tags.