Merge docs job into CI workflow
To match how Safir handles this and so that the PyPI upload can
depend on a successful documentation and Docker build, merge the
docs build into the main CI workflow and remove the path conditions.
Make the upload conditional on a push or a pull request from a
tickets branch to reproduce some of the branch logic from the
original workflow.

Remove the test for github.repository when deciding to do the
Docker build since this doesn't do anything (it's always set to the
local repository).  PRs from untrusted repositories that use our
tickets/ branch naming convention will attempt Docker builds and
doc uploads but will fail, which seems acceptable.